Actions to increase/decrease font size by fixed steps
Issue is that in my setup the bitmap font is not rendered correctly.
Therefore I'd like to configure fixed breakpoints for the font sizes
in addition to what alacritty already provides.

This commit introduces two actions:

Action::IncreaseFontSizeBreakpoints
Action::DecreaseFontSizeBreakpoints

There are hardcoded breakpoints currently, which I find optimal for my
font: 13.0, 19.5, 22.5, 25.5, 32.0 .

Perhaps these breakpoints could be taken from a new option in the
alacritty configuration as the next step.